fix: Undefined peer connection error (#244)
There is an error that appears from time to time in Rollbar:

`Uncaught TypeError: Cannot read property 'setRemoteDescription' of undefined`

It seems it can occur if there is a problem when building the RTCPeerConnection while performing a handshake to establish a WebRTC connection. Not sure what could cause that, but we should handle it using the normal error flow, so maybe the Peer library can try to recover.